Delegate eval directly to Harbor
